feat: upgrade contract to version 2.0.0 using cosmwasm 1.0.0
