+++ contracts/dca/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "astroport-dca-module"
-version = "1.0.0"
+version = "2.0.0"
 authors = ["Astroport", "Kaimen Sano"]
 edition = "2021"
 description = "The Astroport DCA module contract implementation"
@@ -20,19 +20,20 @@ crate-type = ["cdylib", "rlib"]
 [features]
 # for quicker tests, cargo test --lib
 # for more explicit tests, cargo test --features=backtraces
-backtraces = ["cosmwasm-std/backtraces", "cosmwasm-vm/backtraces"]
+backtraces = ["cosmwasm-std/backtraces"]
 
 [dependencies]
-astroport = { git = "https://github.com/astroport-fi/astroport-core.git", branch = "main" }
-astroport-dca = { path = "../../packages/astroport-dca", version = "1.0.0" }
-cw2 = { version = "0.8" }
-cw20 = { version = "0.8" }
-cosmwasm-std = { version = "0.16.2" }
-cw-storage-plus = {version = "0.8.0"}
+astroport = { version = "2.0.0", git = "https://github.com/astroport-fi/astroport-core.git" }
+astroport-dca = { version = "2.0.0", path = "../../packages/astroport-dca" }
+
+cw-storage-plus = "0.14.0"
+cw2 = "0.14.0"
+cw20 = "0.14.0"
+cosmwasm-std = "1.0.0"
+
 schemars = "0.8.1"
 serde = { version = "1.0.103", default-features = false, features = ["derive"] }
 thiserror = { version = "1.0.20" }
 
 [dev-dependencies]
-cosmwasm-vm = { version = "0.16.2", default-features = false, features = ["iterator"] }
-cosmwasm-schema = "0.16.2"
+cosmwasm-schema = "1.0.0"
