+++ packages/astroport-dca/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "astroport-dca"
-version = "1.0.0"
+version = "2.0.0"
 authors = ["Astroport", "Kaimen Sano"]
 edition = "2018"
 repository = "https://github.com/astroport-fi/astroport-dca"
@@ -14,14 +14,8 @@ homepage = "https://astroport.fi"
 backtraces = ["cosmwasm-std/backtraces"]
 
 [dependencies]
-cw20 = { version = "0.8" }
-terra-cosmwasm = { version = "2.2.0" }
-cosmwasm-std = { version = "0.16.0", features = ["iterator"] }
+cw20 = "0.14.0"
+cosmwasm-std = "1.0.0"
 schemars = "0.8.1"
 serde = { version = "1.0.103", default-features = false, features = ["derive"] }
-uint = "0.9.1"
-regex = "1.5"
-astroport = {git = "https://github.com/astroport-fi/astroport-core.git", branch = "main", package = "astroport"}
-
-[dev-dependencies]
-cosmwasm-schema = "0.16.0"
+astroport = { version = "2.0.0", git = "https://github.com/astroport-fi/astroport-core.git" }
