+++ packages/astroport-dca/src/lib.rs
@@ -1,1 +1,7 @@
-pub mod dca;
+mod dca;
+mod msg;
+mod user_config;
+
+pub use dca::DcaInfo;
+pub use msg::*;
+pub use user_config::UserConfig;
